requirements.txt
Django==1.5
wsgiref==0.1.2
django-debug-toolbar==0.9.4
django-compressor==1.2
South==0.7.6

# For examples
django-mptt==0.5.5
Django-Select2==3.2.1

# Sphinx documentation
Sphinx==1.1.3
